+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="ercanduman.fcmmessagingdemo.FirstFragment"
+        android:name="ercanduman.fcmmessagingdemo.ui.main.FirstFragment"
         android:label="@string/first_fragment_label"
         tools:layout="@layout/fragment_first">
 
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/SecondFragment"
-        android:name="ercanduman.fcmmessagingdemo.SecondFragment"
+        android:name="ercanduman.fcmmessagingdemo.ui.main.SecondFragment"
         android:label="@string/second_fragment_label"
         tools:layout="@layout/fragment_second">
